move styled components to external files
