+++ client/src/components/PageLayout.jsx
@@ -7,7 +7,8 @@ import './PageLayout.css';
 import React from 'react';
 import { Outlet  } from "react-router-dom";
 import { ThemeProvider } from '@mui/material/styles';
-import { theme, Sidebar } from './theme.jsx';
+import { theme } from './theme.jsx';
+import { Sidebar } from '../styled-components/Sidebar.js';
 import useMediaQuery from '@mui/material/useMediaQuery';
 import Box from '@mui/material/Box';
 import Button from '@mui/material/Button';
