+++ client/src/components/theme.jsx
@@ -1,34 +1,4 @@
-import { createTheme, styled } from '@mui/material/styles';
-import Drawer from '@mui/material/Drawer';
-import ListItemButton from '@mui/material/ListItemButton';
-import ListItemText from '@mui/material/ListItemText';
-
-export const Sidebar = styled(Drawer)(() => ({
-    '.MuiPaper-root': {
-        backgroundColor: theme.palette.background.sidebar
-    }
-}));
-
-export const StyledListItemButton = styled(ListItemButton)(() => ({
-    paddingLeft: '2px',
-    paddingRight: '8px',
-    borderRadius: '8px',
-    height: '64px',
-    '&.Mui-selected': {
-        backgroundColor: theme.palette.greys.light,
-        ':hover': {
-            backgroundColor: theme.palette.greys.medium
-        }
-    }
-}));
-
-export const StyledListItemText = styled(ListItemText)(() => ({
-    marginLeft: '8px',
-}));
-
-export const listItemTypography = {
-    fontFamily: 'Arial'
-}
+import { createTheme } from '@mui/material/styles';
 
 export const theme = createTheme({
     components: {
