+++ client/src/styled-components/TabListItem.js
@@ -1,0 +1,25 @@
+import {styled} from '@mui/material/styles';
+import {theme} from '../components/theme.jsx';
+import ListItemButton from '@mui/material/ListItemButton';
+import ListItemText from '@mui/material/ListItemText';
+
+export const StyledListItemButton = styled(ListItemButton)(() => ({
+    paddingLeft: '2px',
+    paddingRight: '8px',
+    borderRadius: '8px',
+    height: '64px',
+    '&.Mui-selected': {
+        backgroundColor: theme.palette.greys.light,
+        ':hover': {
+            backgroundColor: theme.palette.greys.medium
+        }
+    }
+}));
+
+export const StyledListItemText = styled(ListItemText)(() => ({
+    marginLeft: '8px',
+}));
+
+export const listItemTypography = {
+    fontFamily: 'Arial'
+}
